backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
asyncpg==0.29.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
requests==2.31.0
boto3==1.34.0
azure-identity==1.15.0
google-cloud-securitycenter==3.6.0
docker==7.0.0
kubernetes==28.1.0
terraform-docs==0.16.0
PyYAML==6.0.1
cryptography==41.0.8
python-dotenv==1.0.0
aiofiles==23.2.1
celery==5.3.4
redis==5.0.1
pygithub==2.1.1
python-gitlab==4.1.0